test(pathfinding): v6 predictive-Z round-trip coverage; drop v5 tests
Predictor lossless round-trip (incl. wraparound extremes) + Predict masking; full-record
identity for flat-elide (asserts compact size), sloped subset, varied, coastline (the
masked-predictor correctness case: nonzero SourceZ, partial walkability), swim-layer,
strata, swim+strata combined, and older-version-rejected. Removes the superseded v5 tests.
